fraud detection APIs implemented

+++ src/main/java/com/bankapplication/service/accountService/AccountServiceImpl.java
@@ -15,6 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -38,7 +39,7 @@ public class AccountServiceImpl implements AccountService {
 
         Account account = new Account();
         account.setAccountType(accountDto.getAccountType());
-        account.setBalance(accountDto.getBalance());
+        account.setBalance(BigDecimal.valueOf(accountDto.getBalance()));
         account.setUser(user);
 
         String accountNumber = AccountNumberGenerator.generateAccountNumber();
